+++ src/gif/block/labeled/Trailer.java
@@ -4,10 +4,10 @@ import java.io.IOException;
 import java.io.OutputStream;
 
 public class Trailer extends LabeledBlock {
-  public static final int label = 0x3b;
+  public static final byte label = 0x3b;
 
   @Override
-  public int getLabel() { return Trailer.label; }
+  public byte getLabel() { return Trailer.label; }
 
   @Override
   public void writeTo(OutputStream stream) throws IOException {
