+++ src/gif/data/format/ByteFormatter.java
@@ -1,7 +1,7 @@
 package gif.data.format;
 
 public class ByteFormatter {
-  public static String format(int x) {
+  public static String format(byte x) {
     return String.format("0x%02x", x);
   }
 }
